feat: return None instead of magic NA values if field is not defined
Code that uses this library shouldn't have to be aware that some return values actually mean that no data was provided.

Caution: This abstraction is slightly leaky. If a user provides one of these magic values and then reads the message back, it will instead return 'None' for the field.